+++ src/sys/hw/mcu/board/_.rs
@@ -18,7 +18,7 @@ crate::mods_in! {
 crate::mods_out! { // _mods
     _mods {
         pub use super::{
-            arduino::_all::ArduinoNano,
+            arduino::_all::BoardArduinoNano,
         };
     }
 }