+++ src/sys/hw/mcu/board/arduino/_.rs
@@ -9,7 +9,7 @@ crate::mods_in! {
 crate::mods_out! { // _mods
     _mods {
         pub use super::{
-            nano::ArduinoNano,
+            nano::BoardArduinoNano,
         };
     }
 }